more validation! TODO: show the user what's wrong

+++ clan/src/de/bastian/clan/server/domain/Picture.java
@@ -14,6 +14,7 @@ import javax.persistence.Version;
 
 import com.google.appengine.api.datastore.Text;
 
+import de.bastian.clan.shared.UserProxy;
 import de.bastian.clan.shared.ValidationException;
 
 @PersistenceCapable(identityType = IdentityType.APPLICATION)
@@ -104,9 +105,11 @@ public class Picture {
             if (getId() == null) { // only on create...
                 setCreated(new Date());
                 setVersion(1);
-            } else {
+            } else if (getUser() == User.isLoggedIn().getId() || User.isLoggedIn().getType().equals(UserProxy.Type.Admin)) {
                 setChanged(new Date());
                 setVersion(getVersion() + 1);
+            } else {
+                throw new ValidationException();
             }
 
             pm.makePersistent(this);
@@ -115,12 +118,15 @@ public class Picture {
         }
     }
 
-    public void remove() {
+    public void remove() throws ValidationException {
         PersistenceManager pm = persistenceManager();
 
         try {
-            Picture picture = pm.getObjectById(Picture.class, getId()); // TODO: handle post.getPost() != null
+            if (User.isLoggedIn() == null || (getUser() != User.isLoggedIn().getId() && User.isLoggedIn().getType().equals(UserProxy.Type.Admin))) {
+                throw new ValidationException();
+            }
 
+            Picture picture = pm.getObjectById(Picture.class, getId());
             pm.deletePersistent(picture);
         } finally {
             pm.close();
